Major step forward. Added code generators to be able to synthesize test cases for model-specific instances. Only visitor is properly working, but this logic will be transfered to the other modes as well.

+++ src/main/java/expression/data/Eval.java
@@ -6,6 +6,6 @@ import expression.types.Types;
 public class Eval extends Operation {
 
     public Eval() {
-        super("eval", Types.Int);
+        super("eval", Types.Double);
     }
 }
